Add transitional model to represent a FeedItem (post) in the API module.
The purpose of this transitional model is to hide API implementation details from the Feed module. For instance, we might have different properties names and different types. In this particular case, the FeedItem model must have a date property of type Date. The mapping from UTC numeric representation of the creation date to the Date type is done inside this new model. This way the FeedItem remains agnostic of API details.